refactor: Carpeta components y view de activity #140
