+++ src/activity/components/createAhorcado/ahorcadoDifficultySelector/AhorcadoDificultySelector.module.css
