+++ src/activity/components/createCompletarOracion/sentenceConfig/SentenceConfig.module.css
