+++ src/activity/components/createCompletarOracion/wordSelector/WordSelector.module.css
