--- src/activity/components/juegoPrueba/JuegoPrueba.tsx
@@ -1,10 +1,0 @@
-const JuegoPrueba = () => {
-  return (
-    <div style={{ display: "grid", placeContent: "center", fontSize: "3rem" }}>
-      {" "}
-      <h1>JuegoPrueba</h1>
-    </div>
-  );
-};
-
-export default JuegoPrueba;
