+++ src/activity/components/preguntados/generalConfig/GeneralConfiguration.module.css
