--- src/activity/contexts/ahorcadoContext/AhoracadoContext.type.ts
@@ -1,8 +1,0 @@
-
-
-import type { AhorcadoInterface } from "../../types/Ahorcado.type";
-
-export interface AhorcadoContextType {
-  loading: boolean;
-  registerAhorcado: (data:AhorcadoInterface) => Promise<void>;
-}
