--- src/activity/contexts/ahorcadoContext/AhorcadoContext.ts
@@ -1,6 +1,0 @@
-import { createContext } from "react";
-import type { AhorcadoContextType } from "./AhoracadoContext.type";
-
-export const AhorcadoContext = createContext<AhorcadoContextType  | undefined>(
-  undefined
-);
