+++ src/activity/contexts/createAhorcadoContext/CreateAhorcadoContext.ts
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+import type { CreateAhorcadoContextType } from "./CreateAhorcadoContext.type";
+
+export const CreateAhorcadoContext = createContext<
+  CreateAhorcadoContextType | undefined
+>(undefined);
